chore: cleanup feature flags, merge taskmanager
