+++ crates/topos-tce-broadcast/Cargo.toml
@@ -31,9 +31,6 @@ rand.workspace = true
 
 topos-test-sdk = { path = "../topos-test-sdk/" }
 
-[features]
-task-manager-channels = []
-
 [[bench]]
 name = "double_echo"
 path = "benches/double_echo.rs"
